min op

+++ 0x02-minimum_operations/0-minoperations.py
@@ -11,7 +11,7 @@ def minOperations(n: int) -> int:
     while n > 1:
         for i in factors:
             if n % i == 0:
-                ops.append(n / i)
+                ops.append(int(n / i))
                 n /= i
                 break
